requirements.txt
python-telegram-bot
Rx
opencv-python
layered-yaml-attrdict-config